“添加Gallery.harmony.js文件
