+++ CHANGES.md
@@ -1,8 +1,9 @@
 
 ## v2.1.5-0.0.1
 
-1.修改源库依赖
-2.添加harmonyOS目录
+1.将ViewPropTypes接口替换为PropTypes
+2.判断UIManager是否存在measure
+3.修改ViewTransformer普通函数改为箭头函数，改变this的指向问题
 
 react-native-openharmony: 0.72.20
 DevEco Studio: 5.0.3.200
